+++ android/src/main/java/com/jiguang/jpush/JPushPlugin.java
@@ -5,6 +5,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Handler;
 import android.os.Looper;
+import android.text.TextUtils;
 import android.util.Log;
 
 import org.json.JSONObject;
@@ -17,7 +18,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import cn.jiguang.api.JCoreInterface;
 import cn.jpush.android.api.JPushInterface;
+import cn.jpush.android.api.NotificationMessage;
 import cn.jpush.android.data.JPushLocalNotification;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.plugin.common.MethodCall;
@@ -110,9 +113,23 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
             isNotificationEnabled(call, result);
         } else if (call.method.equals("openSettingsForNotification")) {
             openSettingsForNotification(call, result);
+        } else if (call.method.equals("setWakeEnable")) {
+            setWakeEnable(call, result);
         } else {
             result.notImplemented();
         }
+    }
+
+    private void setWakeEnable(MethodCall call, Result result) {
+        HashMap<String, Object> map = call.arguments();
+        if (map == null) {
+            return;
+        }
+        Boolean enable = (Boolean) map.get("enable");
+        if (enable == null) {
+            enable = false;
+        }
+        JCoreInterface.setWakeEnable(context,enable);
     }
 
     // 主线程再返回数据
@@ -471,6 +488,53 @@ public class JPushPlugin implements FlutterPlugin, MethodCallHandler {
 
     }
 
+    public static void onNotifyMessageUnShow( NotificationMessage notificationMessage) {
+        Log.e(TAG,"[onNotifyMessageUnShow] message:"+notificationMessage);
+        if (instance == null) {
+            Log.d(TAG, "the instance is null");
+            return;
+        }
+        Map<String, Object> notification= new HashMap<>();
+        notification.put("title", notificationMessage.notificationTitle);
+        notification.put("alert", notificationMessage.notificationContent);
+        notification.put("extras", getExtras(notificationMessage));
+        JPushPlugin.instance.channel.invokeMethod("onNotifyMessageUnShow", notification);
+    }
+
+
+    private static Map<String,Object> getExtras(NotificationMessage notificationMessage){
+        Map<String, Object> extras= new HashMap<>();
+        try {
+            extras.put(JPushInterface.EXTRA_MSG_ID, notificationMessage.msgId);
+            extras.put(JPushInterface.EXTRA_NOTIFICATION_ID, notificationMessage.notificationId);
+            extras.put(JPushInterface.EXTRA_ALERT_TYPE, notificationMessage.notificationAlertType + "");
+            if (!TextUtils.isEmpty(notificationMessage.notificationExtras)) {
+                extras.put(JPushInterface.EXTRA_EXTRA, notificationMessage.notificationExtras);
+            }
+            if (notificationMessage.notificationStyle == 1 && !TextUtils.isEmpty(notificationMessage.notificationBigText)) {
+                extras.put(JPushInterface.EXTRA_BIG_TEXT, notificationMessage.notificationBigText);
+            } else if (notificationMessage.notificationStyle == 2 && !TextUtils.isEmpty(notificationMessage.notificationInbox)) {
+                extras.put(JPushInterface.EXTRA_INBOX, notificationMessage.notificationInbox);
+            } else if ((notificationMessage.notificationStyle == 3) && !TextUtils.isEmpty(notificationMessage.notificationBigPicPath)) {
+                extras.put(JPushInterface.EXTRA_BIG_PIC_PATH, notificationMessage.notificationBigPicPath);
+            }
+            if (!(notificationMessage.notificationPriority == 0)) {
+                extras.put(JPushInterface.EXTRA_NOTI_PRIORITY, notificationMessage.notificationPriority + "");
+            }
+            if (!TextUtils.isEmpty(notificationMessage.notificationCategory)) {
+                extras.put(JPushInterface.EXTRA_NOTI_CATEGORY, notificationMessage.notificationCategory);
+            }
+            if (!TextUtils.isEmpty(notificationMessage.notificationSmallIcon)) {
+                extras.put(JPushInterface.EXTRA_NOTIFICATION_SMALL_ICON, notificationMessage.notificationSmallIcon);
+            }
+            if (!TextUtils.isEmpty(notificationMessage.notificationLargeIcon)) {
+                extras.put(JPushInterface.EXTRA_NOTIFICATION_LARGET_ICON, notificationMessage.notificationLargeIcon);
+            }
+        }catch (Throwable e){
+            Log.e(TAG,"[onNotifyMessageUnShow] e:"+e.getMessage());
+        }
+        return extras;
+    }
     static void transmitNotificationReceive(String title, String alert, Map<String, Object> extras) {
         Log.d(TAG, "transmitNotificationReceive " + "title=" + title + "alert=" + alert + "extras=" + extras);